Fixing spec test memory bounds calculation

+++ Wacs.Core/Instructions/SIMD/VMemory.cs
@@ -72,7 +72,7 @@ namespace Wacs.Core.Instructions.SIMD
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //7.
-            uint i = context.OpStack.PopI32();
+            long i = context.OpStack.PopI32();
             //8.
             long ea = (long)i + (long)M.Offset;
             //9.
@@ -177,7 +177,7 @@ namespace Wacs.Core.Instructions.SIMD
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //7.
-            uint i = context.OpStack.PopI32();
+            long i = context.OpStack.PopI32();
             //8.
             long ea = (long)i + (long)M.Offset;
             //9.
@@ -282,7 +282,7 @@ namespace Wacs.Core.Instructions.SIMD
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //7.
-            uint i = context.OpStack.PopI32();
+            long i = context.OpStack.PopI32();
             //8.
             long ea = (long)i + (long)M.Offset;
             //9.
@@ -390,7 +390,7 @@ namespace Wacs.Core.Instructions.SIMD
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //9.
-            uint i = context.OpStack.PopI32();
+            long i = context.OpStack.PopI32();
             //10.
             long ea = (long)i + (long)M.Offset;
             //11.
@@ -503,9 +503,9 @@ namespace Wacs.Core.Instructions.SIMD
             context.Assert( context.OpStack.Peek().IsI32,
                 $"Instruction {Op.GetMnemonic()} failed. Wrong type on stack.");
             //9.
-            uint i = context.OpStack.PopI32();
+            long i = context.OpStack.PopI32();
             //10.
-            uint ea = i + M.Offset;
+            long ea = i + M.Offset;
             //11.
             if (ea + WidthN.ByteSize() > mem.Data.Length)
                 throw new TrapException($"Instruction {Op.GetMnemonic()} failed. Memory pointer out of bounds.");
